% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_member.R
\docType{data}
\name{list_member}
\alias{list_member}
\title{List member data}
\format{
A tibble with the following variables (types) and definitions:
\describe{
\item{created_at}{(POSIXct) Date and time when the user account was created, in ETC.}
\item{username}{(character) Handle (screen name) the user identifies themselves with. Usernames are unique but subject to change.}
\item{name}{(character) Name of the user, as they chose to represent it on their profile. Not necessarily a person or company’s name.}
\item{description}{(character) User's bio or profile description.}
\item{followers_count}{(integer) Number of users following the user.}
\item{following_count}{(integer) Number of users the user is following.}
\item{listed_count}{(integer) Number of lists the user is listed in.}
\item{tweet_count}{(integer) Number of tweets the user has posted.}
\item{is_verified}{(logical) Indicates if the user is verified.}
\item{verified_type}{(character) Indicates the type of verification for the user (blue, business, government, or none).}
\item{location}{The location specified in the user's profile. As this is a freeform value, it may not indicate a valid location, but it may be fuzzily evaluated when performing searches with location queries.}
\item{profile_image_url}{The URL to the user's profile image.}
\item{list_id}{(character) Unique identifier for the list the acccount is a member of, if any.}
\item{user_id}{(character) Unique identifier for the user.}
}
}
\source{
The \href{https://developer.twitter.com/en/docs/twitter-api/lists/list-members/api-reference/get-lists-id-members}{List Members} and \href{https://developer.twitter.com/en/docs/twitter-api/tweets/timelines/api-reference/get-users-id-tweets}{Timelines} endpoints of the Twitter API.
}
\usage{
list_member
}
\description{
Each row in this table is a member of one of the lists in the \code{list}
table. The data in this table stem from the information you can find on the
user's profile page. It is current as of this package's most recent update.
Therefore, the user's description, verified status, etc., may differ today.
You can join the user information in this table with rows in the \code{list}
and \code{tweet} tables via the \code{list_id} and \code{user_id} variables,
respectively.
}
\examples{
# Display the `list_member` tibble in the console
list_member
}
\keyword{datasets}
